+++ maze.py
@@ -5,6 +5,7 @@ import time
 import os
 from ColorDictionary import ColorDicionary
 from enum import Enum
+from UIUtility import Button, ButtonSlider
 #Classes
 
 # Tank sprite class
@@ -662,137 +663,13 @@ class Explosion(pygame.sprite.Sprite):
     def getCooldown(self):
         return self.animationCooldown
 
-class Button:
-    buttonState = False #False = Not clicked, True = Clicked
-    def __init__(self, color = (0, 0, 0), secondaryColor = (255, 255, 255), x = 0, y = 0, width = 0, height = 0, text = '', textColor = (0, 0, 0)):
-        self.color = color
-        self.x = x
-        self.y = y
-        self.width = width
-        self.height = height
-        self.text = text
-        self.textColor = textColor
-        self.secondaryColor = secondaryColor
-        self.display = self.color
-    def draw(self, screen, outline=None):
-        pygame.draw.rect(screen, self.display, (self.x, self.y, self.width, self.height), 0)
-        if outline:
-            pygame.draw.rect(screen, (0, 0, 0), (self.x, self.y, self.width, self.height), 1)
-
-
-        if self.text != '':
-            font = pygame.font.SysFont('Ariel', 20)
-            text=font.render(self.text, 1, self.textColor)
-            screen.blit(text, (self.x + (self.width/2 - text.get_width()/2), self.y + (self.height/2 - text.get_height()/2)))
-
-    def ButtonClick(self, mouse):
-        if not(self.x < mouse[0] < self.x + self.width and self.y < mouse[1] < self.y + self.height):
-            return # We didn't click this button
-        self.buttonState = not self.buttonState
-        if self.buttonState:
-            self.display = self.secondaryColor
-        else:
-            self.display = self.color
-        pass # Any action that needs to be processed when the button is clicked
-
-    def getCorners(self):
-        return (self.x, self.y, self.x+self.width, self.y+self.height)
-
-class ButtonSlider:
-     
-    carrierX = 20
-    carrierY = 10
-    carLocationX = 0
-    barY = 4
-    buttonSpacing = 50
-    clicked = False
-    def __init__(self, color = (0, 0, 0), secondaryColor = (255, 255, 255), x = 0, y = 0, buttonWidth = 0, buttonHeight = 0, width = 0, height = 0, text = '', textColor = (0, 0, 0), buttonColor = (0,0,0), buttonSecondaryColor=(0,0,0)):
-        self.color = color
-        self.x = x
-        self.y = y
-        self.buttonWidth = buttonWidth
-        self.buttonHeight = buttonHeight
-        self.width = width
-        self.height = height
-        self.secondaryColor = secondaryColor
-        self.carrierX = height//5
-        self.carrierY = height//2
-        self.carLocationX = self.x + self.width
-        self.text = text
-        self.textColor = textColor
-        self.display = self.color
-        self.buttonColor = buttonColor
-        self.buttonSecondaryColor = buttonSecondaryColor
-    def draw(self, screen, outline=None):
-
-        #Button
-        pygame.draw.rect(screen, self.display, (self.x, self.y - self.buttonHeight/2, self.buttonWidth, self.buttonHeight), 0) #Button
-
-        #Text
-        if self.text != '':
-            font = pygame.font.SysFont('Ariel', 20)
-            text=font.render(self.text, 1, self.textColor)
-            screen.blit(text, (self.x + (text.get_width()/2) - self.buttonWidth/4, self.y + (text.get_height()/2)-self.buttonHeight/4))
-
-        #Slider
-        pygame.draw.rect(screen, self.color, (self.x + self.buttonSpacing *2, self.y-self.barY, self.width, self.barY*2), 0) # Bar
-
-        pygame.draw.rect(screen, self.secondaryColor, (self.carLocationX - self.carrierX/2 + self.buttonSpacing *2, self.y - self.carrierY/2, self.carrierX, self.carrierY), 0) # Carrier
-
-        #Percentage bar
-        text = pygame.font.SysFont('Arial', 50).render(str(int(self.getPercentage())), True, (0, 0, 0))
-        screen.blit(text, (self.x + self.width + self.buttonSpacing*2.5, self.y - text.get_height()//2))
-
-        #Outlines
-        if outline:
-            pygame.draw.rect(screen, (0, 0, 255), (self.x + self.buttonSpacing*2, self.y - self.carrierY/2, self.width, self.carrierY), 1)
-
-    def ButtonClick(self):
-        print("Button clicked")
-        self.clicked = not self.clicked
-        if self.clicked:
-            self.display = self.buttonSecondaryColor
-        else:
-            self.display = self.buttonColor
-
-
-    def getCorners(self):
-        return (self.x, self.y - self.buttonHeight/2, self.x+self.buttonWidth, self.y + self.buttonHeight / 2)
-
-    def updateSlider(self, mouseX, mouseY):
-        if not(self.y - self.height/2 < mouseY < self.y + self.height/2 and self.x + self.buttonSpacing*2 < mouseX < self.x + self.width + self.buttonSpacing*2):
-            return
-        #If we are on the slier
-        if self.clicked:
-            self.ButtonClick()
-        self.carLocationX = mouseX - self.buttonSpacing*2
-
-    def getPercentage(self):
-        percentage = round(round((self.carLocationX - self.x) / (self.width), 2)*100,3)
-        return percentage
-
-    def checkButtonClick(self, mouseX, mouseY):
-        if self.x < mouseX < self.x + self.buttonWidth and self.y < mouseY + self.buttonHeight/2 < self.y + self.buttonHeight:
-            self.ButtonClick()
-
-    def getValue(self):
-        if self.clicked:
-            return 0
-        return self.getPercentage()/100
-
-    def mute(self):
-        self.clicked = True
-        self.display = self.buttonSecondaryColor
-        
-
 class GameMode(Enum):
     #This class is responsible for the game mode
     # This class doesn't have other function as they are not needed
-    play = 1
     pause = 0
+    play = 1
     home = 2
     settings = 3
-
 
 #Functions
 def tileGen():
@@ -1063,7 +940,9 @@ def playGame():
     explosionGroup.draw(screen)
 
 def pauseScreen():
-
+    # This function will draw the pause screen
+    # Inputs: None
+    # Outputs: None
     pauseWidth = windowWidth - mazeX * 2
     pauseHeight = windowHeight - mazeY * 2
     pygame.draw.rect(screen, c.geT("OFF_WHITE"), [mazeX, mazeY, pauseWidth, pauseHeight])
@@ -1179,6 +1058,7 @@ indentFromLeft = mazeX
 indentFromRight = mazeY
 sliderX = indentFromLeft * 2.5
 sliderY = windowHeight/8
+#Buttons
 home = Button(c.geT("GREEN"), c.geT("WHITE"),indentFromLeft * 1.5 , indentFromRight * 1.8, tileSize, tileSize, 'Home')
 quitButton = Button(c.geT("GREEN"), c.geT("WHITE"), windowWidth - indentFromLeft * 2.5, indentFromRight * 1.8, tileSize, tileSize, 'Quit')
 unPause = Button(c.geT("GREEN"), c.geT("WHITE"), windowWidth/2 - 200, 0.8 * windowHeight, 400, tileSize, 'Return to Game')
